+++ feature/expenses/src/main/java/com/expenseanalyst/feature/expenses/ui/EditExpenseViewModel.kt
@@ -97,6 +97,7 @@ class EditExpenseViewModel @Inject constructor(
                 selectedTags = expense.tags,
                 selectedAccountId = expense.accountId,
                 rawSmsBody = expense.rawSmsBody,
+                sourceSender = expense.sourceSender,
                 expenseSourceType = expense.sourceType,
                 isReimbursable = expense.isReimbursable
             )